Fix an odd Nuke crash under certain circumstances
A chain of events was causing Nuke to crash. It only ever happened on a non dockable window, when preferences didn't already exist.

I narrowed it down to "AbstractWindow.show()" > "AbstractWindow.loadWindowPosition()" > "AbstractWindow.centreWindow()" > "self.parent()", where querying the parent was crashing.

Literally no idea what's going on, I'm assuming it's a bug specific to Nuke, but I've set it so that it'll fallback to using "getMainWindow()" when loading.