Fix bugs with Substance Designer
Calling windowReady too early will result in the window not showing due to the overridden setVisible method, so a timer was added to the deferred method until Substance Designer gets a better option. The window parent also can't be relied upon when docking is used, so the dock widget is stored in a different attribute.