Fix issue with loading GUI on Nuke startup
With the previous commit, the check on `mainWindow.isVisible()` would always return False on startup as the window hadn't yet loaded. This resulted in Nuke being blocked from loading due to the VFXWindow call to `app.exec_()`

I compared what internal widgets were present and changed the check to use those instead. Not 100% happy with the solution but it seems to work well.