Fix reloading Maya docked window with no coordinates
If a window config doesn't yet contain any coordinates, then it will centre the window.
Attempting to load the same Maya docked window twice in this way will cause what appears to be a race condition with Maya's `workspaceControl`, where whatever happens ends up with triggering a window close.

>>> from vfxwindow import VFXWindow
>>> class MyWindow(VFXWindow):
...     WindowID = 'test'
...     WindowDockable = True
>>> MyWindow.show()
<vfxwindow.maya.gui.WindowClass object>

>>> MyWindow.show()
<vfxwindow.maya.gui.WindowClass object>
Error: Internal C++ object (WindowClass) already deleted.
RuntimeError: Internal C++ object (WindowClass) already deleted.

Creating `MayaWindow.centreWindow` and having it call `AbstractWindow.centreWindow` somehow fixed the issue, but that made no sense so I'd prefer not to rely on that. Instead, I've updated the `loadWindowPosition` method so that it will skip over any docked windows for all applications, since moving is not supported in any.

Despite the tiny amount of code, this bug has been annoying me for over half a year and took hours to solve. I was never able to recreate it as the config files would save and cause the error to disappear.