Improved Nuke docking
In Nuke, anything launched on startup had a different scope, so when using `inspect.stack()[-1][0].f_globals`, it didn't read the application globals, which resulted in being unable to launch any docked window from this code.

Switching this to use `import __main__` has solved the issue, and removes the messy requirement of passing `globals()` to any scripts that handle launching Nuke windows.