Fix Blender detection on portable versions
`sys.executable` returns `blender-4.3.0-windows-x64/4.3/python/bin/python.exe`, which was failing the checks.

Annoyingly there's nothing blender specific aside from the top level folder name (which could be renamed), so this is not a 100% accurate solution. Maybe in the future I could make it scan for the Blender executable.